Remove Akamai from LICENCE
Previously, clauses were added to the licence to allow John Chittum
(@j5awry) to contribute to the project whilst working at Akamai.
During this time, he made no copyrightable code commit and as of
Jan 17th 2020 he will no longer be working for Akamai. As such,
there is no need for Flask-RESTX to continue with the licence
additions.