Update project authors
Change project authorship to "python-restx Authors"
and update LICENCE/setup.py accordingly to reflect
ownership of the new fork. The author_email field
has been removed from setup.py as it is no longer
a required field https://github.com/python/cpython/pull/17388
and setting up a python-restx team email is not a
task we're looking to take on at the moment.