Use Coveralls GitHub Action instead of coveralls CLI
Coveralls CLI requires a GITHUB_REPO_TOKEN env var for authentication.
This is set using GitHub secrets which are *not* available to forked
repos - causing the builds to fail on pull requests. The GitHub action
instead uses the built it GitHub Actions GITHUB_TOKEN which is available
to forked repos.